Break off message support code into msglib
