+++ nat20/link.py
@@ -1,108 +1,18 @@
 """
 Contains the messy details of communicating with a pixels die.
 """
-import abc
 import asyncio
 import collections
-import dataclasses
 import inspect
 import logging
-import struct
-import sys
-from typing import Callable, Optional, Self, Iterable
+from typing import Callable
 
 import bleak
 
 from .constants import CHARI_NOTIFY, CHARI_WRITE
-
+from .msglib import Message, pack, unpack
 
 LOG = logging.getLogger(__name__)
-
-
-_messages = {}
-
-
-def iter_msgs() -> Iterable[type['Message']]:
-    """
-    List all known messages.
-    """
-    yield from _messages.values()
-
-
-def msgid(msg) -> Optional[int]:
-    """
-    Gets the ID of the given message.
-
-    Returns :const:`None` if the message doesn't have one.
-    """
-    return msg._Message__id
-
-
-class Message(abc.ABC):
-    """
-    Base class for messages that get communicated with the die.
-
-    Messages must be defined as::
-
-        class Spam(Message, id=42):
-
-    Pass :const:`None` as the ID if this should not be registered
-    as a message (eg, is abstract).
-
-    Args:
-        id (int|None): The message ID or None.
-    """
-    @classmethod
-    @abc.abstractmethod
-    def __struct_unpack__(cls, blob: bytes) -> Self:
-        """
-        Construct an instance from a message blob.
-        """
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def __struct_pack__(self) -> bytes:
-        """
-        Turn this message back into a blob.
-        """
-        raise NotImplementedError
-
-    def __init_subclass__(cls, /, id: int, **kwargs):
-        super().__init_subclass__(**kwargs)
-        if id is not None:
-            cls.__id = id
-            _messages[id] = cls
-
-
-class BasicMessage(Message, id=None):
-    """
-    Provides a helpful basic version of :class:`Message`
-    for standard use cases.
-
-    ::
-
-        @dataclass
-        class Spam(BasicMessage, id=42, format='i'):
-           eggs: int
-
-    Subclasses must also be a :mod:`dataclass <dataclasses>`.
-
-    Args:
-        id (int|None): The message ID, or None
-        format (str): The format in :mod:`struct` form.
-    """
-    def __init_subclass__(cls, /, format: str, **kwargs):
-        super().__init_subclass__(**kwargs)
-        cls.__struct_format = sys.intern(f"<{format}")
-
-    @classmethod
-    def __struct_unpack__(cls, blob: bytes) -> Self:
-        fields = struct.unpack(cls.__struct_format, blob)
-        return cls(*fields)
-
-    def __struct_pack__(self) -> bytes:
-        fields = dataclasses.astuple(self)
-        return struct.pack(self.__struct_format, *fields)
 
 
 def _call_or_task(func, *pargs, **kwargs):
@@ -174,18 +84,12 @@ class PixelLink:
 
         :meta private:
         """
-        msgid, blob = packet[0], packet[1:]
         try:
-            msgcls = _messages[msgid]
-        except KeyError:
-            LOG.error("Unknown message ID=%i", msgid)
+            msg = unpack(packet)
+        except Exception:
+            LOG.exception("Unable to unpack packet %r", packet)
         else:
-            try:
-                msg = msgcls.__struct_unpack__(blob)
-            except Exception:
-                LOG.exception("Problem unpacking packet")
-            else:
-                self._dispatch(msg)
+            self._dispatch(msg)
 
     def _dispatch(self, message: Message):
         """
@@ -208,7 +112,7 @@ class PixelLink:
 
         :meta public:
         """
-        blob = bytes([msgid(message)]) + message.__struct_pack__()
+        blob = pack(message)
         await self._client.write_gatt_char(CHARI_WRITE, blob)
 
     async def _wait(self, msgcls: type[Message]) -> Message:
