+++ pyproject.toml
@@ -12,6 +12,7 @@ packages = [{include = "nat20"}]
 [tool.poetry.dependencies]
 python = "^3.11"
 bleak = "^0.20.2"
+typing-extensions = "^4.6.0"
 
 
 [tool.poetry.group.tui.dependencies]
